+++ domain/src/main/java/com/paya/domain/models/repo/CircleChartDataRepoModel.kt
@@ -1,7 +1,7 @@
 package com.paya.domain.models.repo
 
 data class CircleChartDataRepoModel (
-	val buyPrice: Long,
+	val buyPrice: Float,
 	val quantity: Long,
 	val name: String
 )
